% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm2flux_acc_GWP.R
\name{acc2GWP}
\alias{acc2GWP}
\title{Calculates Global Warming Potential.}
\usage{
acc2GWP(data, CH4_eq = 27, N2O_eq = 273)
}
\arguments{
\item{data}{A data frame containing cumulative emissions data in \eqn{kg \cdot ha^{-1}}.}

\item{CH4_eq}{\eqn{CO_2} equivalents with a time horizon of 100 years for \eqn{CH_4}. Default is 27 \eqn{CO_{2}} equivalents according to IPCC, 2021.}

\item{N2O_eq}{\eqn{CO_2} equivalents on a 100-years horizon for \eqn{N_2O}. Default is 273 \eqn{CO_{2}} equivalents according to IPCC, 2021.}
}
\value{
Data frame containing calculated Global Warming potential in \eqn{CO_{2}} equivalents.
}
\description{
Global Warming potential in \eqn{CO_2} equivalents is calculated from cumulative emissions data in \eqn{kg \cdot ha^{-1}}.
}
\examples{
GWP_test1 <- acc2GWP(acc_flux_df_testA)
GWP_test2 <- acc2GWP(acc_flux_df_testM)
}
