% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak2ppm.R
\name{calibration_curves}
\alias{calibration_curves}
\title{Creates calibration curves using gas standards of known concentrations.}
\usage{
calibration_curves(data, r_square_threshold = 0.9)
}
\arguments{
\item{standard_data}{A data frame with peak area values of gas standards that are sorted by gas chromatography runs}
}
\value{
Data frame containing parameters of the calibration curve
}
\description{
Creates calibration curves using gas standards of known concentrations.
}
\examples{
calibration_results <- calibration_curves(standard_data)
}
