% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm2flux_acc_GWP.R
\name{flux2acc}
\alias{flux2acc}
\title{Converts flux into cumulative emission.}
\usage{
flux2acc(data)
}
\arguments{
\item{data}{A data frame containing gas fluxes for a series of sampling events.}
}
\value{
Data frames containing calculated daily fluxes in \eqn{mg \cdot m^{-2} \cdot h^{-1}} (name: "daily_data") and cumulative emissions \eqn{kg \cdot ha^{-1}} (name: "acc_data").
}
\description{
Interpolates linearly throughout a time series of flux data to calculate daily flux and cumulative emissions.
}
\examples{
# Test 1: Using as input a data rame with flux corrections:
flux2acc(flux_df_testA) # ouputs: acc_flux_df_testA (cumulative emissions) and daily_flux_df_testA (linear flux interpolation).

# Test 2: Using as input a data frame without flux corrections:
flux2acc(flux_df_testM) # ouputs: acc_flux_df_testM (cumulative emissions) and daily_flux_df_testM (linear flux interpolation).
}
