% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm2flux.R
\name{ppm2flux}
\alias{ppm2flux}
\title{Converts gas concentration data (ppm) to flux measurements.}
\usage{
ppm2flux(
  data,
  Timesteps = 4,
  CH4_mass = 0,
  N2O_mass = 0,
  CO2_mass = 0,
  Gas1_mass = 0,
  Gas2_mass = 0,
  Gas3_mass = 0,
  Diagnostics = FALSE,
  R2_Threshold = 0.7,
  Neg_Rate = TRUE
)
}
\arguments{
\item{data}{A data frame containing gas concentration readings over time.}

\item{Timesteps}{Integer indicating the number of time intervals per sampling event.}

\item{Diagnostics}{Logic indicating the need of geneerating diagnostic plots for each individual chamber sampling showing calculated rates for the complete and four alternative models. Designed only for cases in which Timesteps = 4.}

\item{R2_Threshold}{Numeric vector defining an \eqn{R^2} threshold for model fitting  under which corrected fluxes (slopes) from linear models are considered as 0 (no emissions). If the user decides to omit this parameter for flux corrections then it must be changed to 0. If increased, then the more strict the correction.}

\item{Neg_Rate}{Logic activating additional restriction to the flux correction excluding alternative models that result in negative flux (even if resulting \eqn{R^2} is higher than that of the original model). Default is TRUE (does not consider this restriction and negative fluxes are accepted). Note: If complete models result in negative flux (and \eqn{R^2} is higher than the defined R2_Threshold) then the selected model will still be this complete model with negative flux.}

\item{GHG_mass}{Numeric vector specifying molecular weights (g/mol) for target gases (i.e., CH4_mass for \eqn{CH_4}, N2O_mass for \eqn{N_2O}, CO2_mass for \eqn{CO_2}, and Gas1_mass, Gas2_mass, and Gas3_mass for Gas1, Gas2 and Gas3, respectively).}
}
\value{
Data frame containing calculated Gas fluxes in \eqn{mg \cdot m^{-2} \cdot h^{-1}}.
}
\description{
Fits linear models to input gas concentration data in ppm per sampling time in minutes (Timesteps) and calculates gas fluxes in mg m\if{html}{\out{<sup>}}-2\if{html}{\out{</sup>}} h\if{html}{\out{<sup>}}-1\if{html}{\out{</sup>}} from model slopes.
}
\examples{
# Load test data set 1
data(input_test1)
flux_df_testA <- ppm2flux(input_test1, CH4_mass = 16, N2O_mass = 44) # e.g. 1: Keeping all arguments as default - Output with alternative models but no diagnostics.
flux_df_testB <- ppm2flux(input_test1, CH4_mass = 16, N2O_mass = 44, Diagnostics = TRUE) # e.g. 2: Activating diagnostic plots - Output with alternative models and diagnostics.
flux_df_testC <- ppm2flux(input_test1, CH4_mass = 16, N2O_mass = 44, Timesteps = 5) # e.g. 3: Testing Timesteps - Output without alternative models nor diagnostics.
flux_df_testD <- ppm2flux(input_test1, CH4_mass = 16, N2O_mass = 44, Diagnostics = TRUE, Timesteps = 5) # e.g. 4: No Diagnostics and Timesteps conflicts - Output without alternative models nor diagnostics.
flux_df_testE <- ppm2flux(input_test1, CH4_mass = 16, N2O_mass = 44, Neg_Rate = FALSE) # e.g. 5: Testing Neg_Rate argument, does not consider alternative models resulting in negative rates.
flux_df_testF <- ppm2flux(input_test1, CH4_mass = 16, N2O_mass = 44, Diagnostics = TRUE, Neg_Rate = FALSE) # e.g. 6: Tests Diagnostic plots with Neg_Rate modified to FALSE.

# Load test data set 2
data(input_test2)
flux_df_testG <- ppm2flux(input_test2, CH4_mass = 16, N2O_mass = 44) # e.g. 7: 2nd test data set. Keeping all arguments as default - Output with alternative models but no diagnostics.
flux_df_testH <- ppm2flux(input_test2, CH4_mass = 16, N2O_mass = 44, Diagnostics = TRUE) # e.g. 8: Activating diagnostic plots - Output with alternative models and diagnostics.

# Load test data set 3
data(input_test3)
flux_df_testI <- ppm2flux(input_test3, CH4_mass = 16, N2O_mass = 44) # e.g. 9: Keeping all arguments as default - Output with alternative models but no diagnostics.

# Load test data set 4
data(input_test4) # 5 timesteps; 1 sampling event
flux_df_testJ <- ppm2flux(input_test4, CH4_mass = 16, N2O_mass = 44, CO2_mass = 44, Timesteps = 5) # e.g. 10: 4th test data set (with 5 timesteps). Keeping all arguments as default - Output without alternative models nor diagnostics.
flux_df_testK <- ppm2flux(input_test4, CH4_mass = 16, N2O_mass = 44, CO2_mass = 44, Diagnostics = TRUE) # e.g. 11: Activating diagnostic plots - Output without alternative models nor diagnostics (due to 5 timesteps within input data frame).

# Load test data set 5
data(input_test5) # 5 timesteps; 16 sampling events
flux_df_testL <- ppm2flux(input_test5, CH4_mass = 16, N2O_mass = 44, CO2_mass = 44, Timesteps = 5) # e.g. 12: 4th test data set (with 5 timesteps). Keeping all arguments as default - Output without alternative models nor diagnostics.
flux_df_testM <- ppm2flux(input_test5, CH4_mass = 16, N2O_mass = 44, CO2_mass = 44, Diagnostics = TRUE) # e.g. 13: Activating diagnostic plots - Output without alternative models nor diagnostics (due to 5 timesteps within input data frame)
}
