% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak2ppm.R
\name{peak2ppm}
\alias{peak2ppm}
\title{Converts gas chromatograph peak area output to concentration}
\usage{
peak2ppm(standard_data, sample_data, r_square_threshold = 0.9)
}
\arguments{
\item{standard_data}{A data frame with peak area values of gas standards that are sorted by gas chromatography runs}

\item{sample_data}{A data frame with peak area values of gas samples that are sorted by gas chromatography runs}

\item{r_square_threshold}{Integer. Defines the r2 values the user would like for the calibration curve generated by gas standards}
}
\value{
Data frame containing calculated sample gas concentrations (ppm)
}
\description{
Creates calibration curves using gas standards per GC run in order to determine sample concentration values
}
\examples{
final_data <- peak2ppm(standard_data, sample_data)
}
